backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chaintale-backend"
version = "0.1.0"
description = "FastAPI backend for the ChainTale collaborative storytelling platform"
requires-python = ">=3.11"
authors = [
    { name = "ChainTale Team" }
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "motor>=3.4.0",
    "pydantic-settings>=2.2.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "argon2-cffi>=23.1.0",
    "python-dotenv>=1.0.1",
    "apscheduler>=3.10.4",
    "httpx>=0.27.0",
    "redis>=5.0.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.5",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
    "types-redis",
    "types-python-jose",
    "types-passlib"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

